Moving database connection to keystone.openDatabaseConnection
