+++ index.js
@@ -114,6 +114,7 @@ Keystone.prototype.initMongo = require('./lib/core/initMongo');
 Keystone.prototype.initNav = require('./lib/core/initNav');
 Keystone.prototype.list = require('./lib/core/list');
 Keystone.prototype.mount = require('./lib/core/mount');
+Keystone.prototype.openDatabaseConnection = require('./lib/core/openDatabaseConnection');
 Keystone.prototype.populateRelated = require('./lib/core/populateRelated');
 Keystone.prototype.redirect = require('./lib/core/redirect');
 Keystone.prototype.render = require('./lib/core/render');
