+++ lib/core/mount.js
@@ -193,76 +193,8 @@ function mount(events) {
 	require('../../server/bindRedirectsHandler')(this, app);
 	require('../../server/bindErrorHandlers')(this, app);
 
-	// Connect to database
-
-	var mongoConnectionOpen = false;
-
-	// support replica sets for mongoose
-	if (this.get('mongo replica set')){
-		debug('setting up mongo replica set');
-		var replicaData = this.get('mongo replica set');
-		var replica = '';
-
-		var credentials = (replicaData.username && replicaData.password) ? replicaData.username + ':' + replicaData.password + '@' : '';
-
-		replicaData.db.servers.forEach(function (server) {
-			replica += 'mongodb://' + credentials + server.host + ':' + server.port + '/' + replicaData.db.name + ',';
-		});
-
-		var options = {
-			auth: { authSource: replicaData.authSource },
-			replset: {
-				rs_name: replicaData.db.replicaSetOptions.rs_name,
-				readPreference: replicaData.db.replicaSetOptions.readPreference
-			}
-		};
-
-		debug('connecting to replicate set');
-		this.mongoose.connect(replica, options);
-
-	} else {
-		debug('connecting to mongo');
-		this.mongoose.connect(this.get('mongo'));
-	}
-
-	this.mongoose.connection.on('error', function(err) {
-
-		if (keystone.get('logger')) {
-			console.log('------------------------------------------------');
-			console.log('Mongo Error:\n');
-			console.log(err);
-		}
-
-		if (mongoConnectionOpen) {
-			if (err.name === 'ValidationError') return;
-			throw err;
-		} else {
-			throw new Error('KeystoneJS (' + keystone.get('name') + ') failed to start - Check that you are running `mongod` in a separate process.');
-		}
-
-	}).on('open', function() {
-
-		debug('mongo connection open');
-		mongoConnectionOpen = true;
-
-		var mounted = events.onMount || function () {};
-
-		var connected = function() {
-			if (keystone.get('auto update')) {
-				debug('applying auto update');
-				keystone.applyUpdates(mounted);
-			} else {
-				mounted();
-			}
-		};
-
-		if (this.sessionStorePromise) {
-			this.sessionStorePromise.then(connected);
-		} else {
-			connected();
-		}
-
-	});
+	// Open database connection
+	this.openDatabaseConnection(events);
 
 }
